rename main to home and deleted home.js file that was'nt being used
closes #33
