+++ routes/displayEntries.js
@@ -4,7 +4,6 @@ function get(request, response) {
     db.query(`select parks.park_name, parks.location, park_comments.text_content from parks 
       left join park_comments on parks.id = park_comments.park_id`).then((result) => {
       const parks = result.rows;
-      console.log(parks)
       let parkList = "";
       for (const park of parks) {
           const {id, park_name, location, text_content} = park;
